requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
astrapy==0.6.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
boto3==1.28.83
botocore==1.31.83
cachetools==5.3.2
cassandra-driver==3.29.0
cassio==0.1.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cohere==4.39
colorama==0.4.6
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.8.0
dj-database-url==2.1.0
Django==5.0
django-environ==0.11.2
django-storages==1.14.2
djangorestframework==3.14.0
fastavro==1.9.2
frozenlist==1.4.1
fsspec==2023.12.2
geomet==0.2.1.post1
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==6.11.0
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.352
langchain-community==0.0.6
langchain-core==0.1.3
langchain-google-genai==0.0.5
langsmith==0.0.74
llama-index==0.9.21
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.2
openai==1.6.0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
proto-plus==1.23.0
protobuf==4.25.1
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rsa==4.9
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
